RavenDB-27269 Move the tests into SlowTests
